improve line selection to be more flexible
This PR makes selecting structs for the `-line` mode flexible. You don't
have to be strict now. If the selection contains any kind of struct,
it'll work. So one selection part can be outside the struct, however if
the other part overlaps any struct selection it's still valid.

This is a backwards compatible change. It improves the overall user
experience and makes it easier for people using with editors.